% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMPTYremove.R
\name{EMPTYremove}
\alias{EMPTYremove}
\title{removes any blank rows and columns from data}
\usage{
EMPTYremove(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame with all the empty rows and columns removed
}
\description{
removes any blank rows and columns from data
}
\examples{
df <- data.frame(a = c(1, NA, 3), b = c(NA, 5, 6))
EMPTYremove(df)

}
