% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAfix.R
\name{NAfix}
\alias{NAfix}
\title{changes values of "N/A" in character data type columns to regular NA data}
\usage{
NAfix(df, string = "N/A")
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame with "N/A" values replaced with NA.
}
\description{
changes values of "N/A" in character data type columns to regular NA data
}
\examples{
df <- data.frame(a = c(1, NA, 3), b = c(NA, 5, 6))
NAfix(df)

}
