% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAreplace.R
\name{NAreplace}
\alias{NAreplace}
\title{Replaces `NA` values in specified numeric columns of a data frame with the column mean or median.}
\usage{
NAreplace(df, columns = NULL, method = "mean")
}
\arguments{
\item{df}{A data frame.}

\item{columns}{all columns named are processed. If left NULL, all columns are used.}

\item{method}{the method used to fill NA values: "mean" or "median".}
}
\value{
A data frame with NA values replaced in the specified columns.
}
\description{
Replaces `NA` values in specified numeric columns of a data frame with the column mean or median.
}
\examples{
df <- data.frame(a = c(1, NA, 3), b = c(NA, 5, 6))
NAreplace(df, columns = c("a"), method = "mean")

}
