% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quicknumeric.R
\name{Quicknumeric}
\alias{Quicknumeric}
\title{converts character data that should be numeric into numeric.}
\usage{
Quicknumeric(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame with the proper rows as the numeric data type
}
\description{
converts character data that should be numeric into numeric.
}
\examples{
df <- data.frame(a = c(1, NA, 3), b = c(NA, 5, 6))
Quicknumeric(df)

}
