% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDnames.R
\name{SDnames}
\alias{SDnames}
\title{standardizes column names by making all characters lowercase, turning spaces and special characters into underscores, and removing extraneous underscores}
\usage{
SDnames(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame with the changes specified above.
}
\description{
standardizes column names by making all characters lowercase, turning spaces and special characters into underscores, and removing extraneous underscores
}
\examples{
df <- data.frame(a = c(1, NA, 3), b = c(NA, 5, 6))
SDnames(df)

}
